requirements.txt
flask
requests
pandas
bokeh
pandas_bokeh
paho-mqtt